# src/routers/userRouter.js
const express = require('express')
const router = new express.Router()
const auth = require('../middleware/auth')
const multer = require('multer')
const sharp = require('sharp')
const {sendWelcomeEmail,sendCancellationEmail}  = require('../emails/account')
const upload = multer({
    limits: {
        fileSize: 2000000
    },
    fileFilter(req, file, callback) {        

        if (!file.originalname.match(/\.(jpg|jpeg|png)$/)) {

            return callback(new Error('pls upload an image'))

        }
        callback(undefined, true)
    }
})

const User = require('../models/user')

router.post('/users', async (req, res) => {
    const user = new User(req.body)

    if (await User.doesExist(user.email)) {
        return res.status(400).send('user already exists')
    }


    try {
        await user.save()
        sendWelcomeEmail(user.email, user.name)
        const token = user.getToken()

        res.status(201).send({ user, token })
    } catch (error) {

        res.status(400).send(error)
    }

})



router.get('/users/me', auth, async (req, res) => {
    res.send(req.user)

})

router.get('/users/login', async (req, res) => {
    try {

        const credentials = Object.keys(req.body)
        const credentialsKey = ['email', 'password']

        const isValidLogin = credentials.every((key) => credentialsKey.includes(key))

        if (!isValidLogin) return res.status(400).send("invalid credentials")

        const user = await User.findByCredentials(req.body.email, req.body.password)

        const token = await user.getToken()

        res.send({ user, token })

    } catch (error) {
        console.log(error);
        res.status(400).send(error)
    }

})

router.post('/users/logout', auth, async (req, res) => {

    try {

        req.user.tokens = req.user.tokens.filter((token) => token.token !== req.token)       
        await req.user.save()
        res.send()

    } catch (error) {
        res.status(500).send()
    }
})

router.post('/users/logoutAll', auth, async (req, res) => {

    try {

        req.user.tokens = []
        await req.user.save()
        res.send()

    } catch (error) {
        res.status(500).send(error)
    }
})

router.delete('/users/me', auth, async (req, res) => {


    try {
        sendCancellationEmail(req.user.email,req.user.name)
        await req.user.remove()        
        res.send(req.user)

    } catch (error) {
        res.status(500).send(error)

    }

})


router.patch('/users/me', auth, async (req, res) => {
    const updates = Object.keys(req.body)
    const allowedUpdates = ['name', 'email', 'password', 'age']

    const isValidUpdate = updates.every((update) => allowedUpdates.includes(update))

    if (!isValidUpdate) {
        return res.status(400).send("Invalid update")

    }

    try {

        updates.forEach((update) => {
            req.user[update] = req.body[update]
        })

        await req.user.save()

        res.send(req.user)

    } catch (error) {
        res.status(500).send(e)

    }

})


router.post('/users/me/avatar', auth, upload.single('avatar'), async (req, res) => {
    
    const imgBuffer = await sharp(req.file.buffer).resize({ width: 250, height: 250 }).png().toBuffer()
    req.user.avatar = imgBuffer
    await req.user.save()
    res.send()
}, (error, req, res, next) => {
    res.status(400).send({ error: error.message })
})


router.delete('/users/me/avatar', auth, async (req, res) => {
    req.user.avatar = undefined
    await req.user.save()
    res.send()
}, (error, req, res, next) => {
    res.status(400).send({ error: error.message })
})

router.get('/users/:id/avatar', async (req, res) => {

    try {
        const user = await User.findById(req.params.id)

        if (!user || !user.avatar) {
            throw new Error()
        }

        res.set('Content-Type', 'image/png')
        res.send(user.avatar)
    } catch (e) {
        res.status(404).send()
    }
})

module.exports = router